+++ lib/src/models/response.dart
@@ -1,23 +1,15 @@
-import 'package:apex_api/src/exceptions/server_exception.dart';
-
 import '../../apex_api.dart';
 
-enum LoginStep {
-  showUsername,
-  showOtp,
-  showPassword,
-  showUpgrade,
-  failure,
-  success
-}
+enum LoginStep { showUsername, showOtp, showPassword, showUpgrade, failure, success }
 
-abstract class BaseResponse {
+class BaseResponse<DM> {
   final int success;
   final String? message;
   Json? data;
   final ServerException? error;
   final String? errorMessage;
   final int expiresAt;
+  final DM? model;
 
   bool get isSuccessful => hasData && success == 1;
 
@@ -44,20 +36,24 @@ abstract class BaseResponse {
     }
   }
 
-  BaseResponse({
-    required this.data,
-    this.error,
-    this.errorMessage,
-  })
+  BaseResponse({this.data, this.error, this.errorMessage, this.model})
       : success = JsonChecker.optInt(data, 'success', defValue: -1)!,
         message = JsonChecker.optString(data, 'message', defValue: null),
-        expiresAt = DateTime
-            .now()
-            .millisecondsSinceEpoch +
-            Duration(
-                seconds: JsonChecker.optInt(data, 'save_local_duration',
-                    defValue: 0)!)
-                .inMilliseconds;
+        expiresAt = DateTime.now().millisecondsSinceEpoch +
+            Duration(seconds: JsonChecker.optInt(data, 'save_local_duration', defValue: 0)!)
+                .inMilliseconds {
+    if (data != null && data!.containsKey('is_logged_in')) {
+      bool saveLocal = JsonChecker.optInt(data, 'save_local_duration', defValue: 0)! > 0;
+      if (!saveLocal && data!['is_logged_in'] == 0) {
+        ApexApiDb.removeToken();
+        // TODO : clear database notify user
+      }
+    }
+
+    if (data != null && data!.containsKey('token') && data!['token'] != null) {
+      ApexApiDb.setToken(data!['token']);
+    }
+  }
 
   bool containsKey(String key) {
     if (data != null) {
@@ -74,45 +70,53 @@ abstract class BaseResponse {
   }
 }
 
-class Response extends BaseResponse {
-  final bool playSound;
+class DataModel {
+  DataModel();
 
-  Response(Map<String, dynamic>? data, {super.error, super.errorMessage})
-      : playSound = data != null && data.containsKey('play_sound')
-      ? (data['play_sound'] == 1)
-      : false,
-        super(data: data) {
-    if (data != null && data.containsKey('is_logged_in')) {
-      bool saveLocal = JsonChecker.optInt(data, 'save_local_duration',
-          defValue: 0)! > 0;
-      if (!saveLocal && data['is_logged_in'] == 0) {
-        ApexApiDb.removeToken();
-        // TODO : clear database notify user
-      }
-    }
-
-    if (data != null && data.containsKey('token') && data['token'] != null) {
-      ApexApiDb.setToken(data['token']);
-    }
-
-    if (message != null) {
-      // TODO : show message anyway
-      // DialogUtil.showNotification(
-      //     success == 1 ? NotificationType.success : NotificationType.error,
-      //     message ?? '');
-    }
+  factory DataModel.fromJson(Json json) {
+    return DataModel();
   }
-
-  factory Response.fromJson(Map<String, dynamic>? json) {
-    if (json != null) return Response(json);
-
-    return Response({
-      'success': -1,
-      'message':
-      'Contact our Tech support @ApexTeamSupport and Ask for help.\nError code : 0x00000000'
-    });
-  }
-
-  @override
-  String toString() => data.toString();
 }
+
+// class Response extends BaseResponse {
+//   final bool playSound;
+//
+//   Response(Map<String, dynamic>? data, {super.model, super.error, super.errorMessage})
+//       : playSound = data != null && data.containsKey('play_sound')
+//       ? (data['play_sound'] == 1)
+//       : false,
+//         super(data: data) {
+//     if (data != null && data.containsKey('is_logged_in')) {
+//       bool saveLocal = JsonChecker.optInt(data, 'save_local_duration',
+//           defValue: 0)! > 0;
+//       if (!saveLocal && data['is_logged_in'] == 0) {
+//         ApexApiDb.removeToken();
+//         // TODO : clear database notify user
+//       }
+//     }
+//
+//     if (data != null && data.containsKey('token') && data['token'] != null) {
+//       ApexApiDb.setToken(data['token']);
+//     }
+//     //
+//     // if (message != null) {
+//     //   // TODO : show message anyway
+//     //   // DialogUtil.showNotification(
+//     //   //     success == 1 ? NotificationType.success : NotificationType.error,
+//     //   //     message ?? '');
+//     // }
+//   }
+//
+//   factory Response.fromJson(Map<String, dynamic>? json, Map<Type, DataModel> responseModels) {
+//     if (json != null) return Response(json, responseModels[Type]);
+//
+//     return Response({
+//       'success': -1,
+//       'message':
+//       'Contact our Tech support @ApexTeamSupport and Ask for help.\nError code : 0x00000000'
+//     });
+//   }
+//
+//   @override
+//   String toString() => data.toString();
+// }