+++ lib/src/socket_join_controller.dart
@@ -1,7 +1,7 @@
 import 'models/response.dart';
 import 'notifier/message_notifier.dart';
 
-class SocketJoinController<T extends Response> extends MessageNotifier {
+class SocketJoinController<T extends BaseResponse<DataModel>> extends MessageNotifier {
   final String groupName;
   final void Function(T data) _onListen;