+++ lib/src/widgets/reactive_widget.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:apex_api/apex_api.dart';
 import 'package:flutter/material.dart';
 
-class ReactiveResponse<Res extends Response> {
+class ReactiveResponse<DM extends DataModel> {
   final ReactiveState state;
-  final Res? response;
+  final BaseResponse<DM>? response;
 
   ReactiveResponse(this.state, {this.response});
 }
@@ -19,22 +19,42 @@ class ReactiveError {
 
 enum ReactiveState { loading, failure, success, error }
 
-class ReactiveController extends ChangeNotifier {
-  void reload() {
-    notifyListeners();
+class ReactiveController<DM extends DataModel> {
+  Future<BaseResponse<DM>> Function()? listener;
+
+  bool get hasClient => listener != null;
+
+  void setListener(Future<BaseResponse<DM>> Function() l) {
+    listener = l;
+  }
+
+  void removeListener() {
+    listener = null;
+  }
+
+  Future<BaseResponse<DM>> reload() {
+    if (listener != null) {
+      return listener!();
+    }
+    return Future.error(Exception('Could not find any listener!'));
   }
 }
 
-class ReactiveWidget<Res extends Response> extends StatefulWidget {
+class ReactiveWidget<DM extends DataModel> extends StatefulWidget {
   final Request request;
   final Widget loadingWidget;
-  final Widget Function(Res response, VoidCallback onRetry) failureWidget;
-  final Widget Function(Res response, VoidCallback onRetry) successWidget;
-  final Widget Function(ServerException exception, Object error, VoidCallback onRetry) retryWidget;
+  final Widget Function(BaseResponse<DM> response, Future<BaseResponse<DM>> Function() onRetry)
+      failureWidget;
+  final Widget Function(BaseResponse<DM> response, Future<BaseResponse<DM>> Function() onRetry)
+      successWidget;
+  final Widget Function(
+          ServerException exception, Object error, Future<BaseResponse<DM>> Function() onRetry)
+      retryWidget;
   final bool ignoreExpireTime;
-  final ReactiveController? controller;
-  final void Function(ReactiveState state, VoidCallback onRetry,
-      {Res? response, ReactiveError? error})? listener;
+  final ReactiveController<DM>? controller;
+  final void Function(ReactiveState state, Future<BaseResponse<DM>> Function() onRetry,
+      {BaseResponse<DM>? response, ReactiveError? error})? listener;
+  final Widget Function(ReactiveState state, Widget child)? wrapper;
 
   const ReactiveWidget({
     Key? key,
@@ -46,108 +66,113 @@ class ReactiveWidget<Res extends Response> extends StatefulWidget {
     this.listener,
     this.ignoreExpireTime = false,
     this.controller,
+    this.wrapper,
   }) : super(key: key);
 
   @override
-  State<ReactiveWidget> createState() => _ReactiveWidgetState<Res>();
+  State<ReactiveWidget> createState() => _ReactiveWidgetState<DM>();
 }
 
-class _ReactiveWidgetState<Res extends Response> extends State<ReactiveWidget<Res>>
-    with WidgetLoadMixin {
-  final StreamController<ReactiveResponse<Res>> _controller =
-      StreamController<ReactiveResponse<Res>>();
+class _ReactiveWidgetState<DM extends DataModel> extends State<ReactiveWidget<DM>>
+    with AutomaticKeepAliveClientMixin {
+  final StreamController<ReactiveResponse<DM>> _controller =
+      StreamController<ReactiveResponse<DM>>();
 
   @override
   void initState() {
     if (widget.controller != null) {
-      widget.controller!.addListener(_sendRequest);
+      widget.controller!.setListener(_sendRequest);
     }
+    _sendRequest();
     super.initState();
   }
 
   @override
   void dispose() {
     if (widget.controller != null) {
-      widget.controller!.removeListener(_sendRequest);
+      widget.controller!.removeListener();
     }
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder<ReactiveResponse<Res>>(
+    return StreamBuilder<ReactiveResponse<DM>>(
       stream: _controller.stream,
       builder: (context, snapshot) {
-        if (snapshot.hasData && snapshot.data != null) {
-          print('hasData');
-          final data = snapshot.data!;
-          if (data.state == ReactiveState.loading) {
-            print('loading');
-            return widget.loadingWidget;
-          } else if (data.state == ReactiveState.failure) {
-            print('failure');
-            return widget.failureWidget(data.response!, _sendRequest);
-          } else if (data.state == ReactiveState.success) {
-            print('success');
-            return widget.successWidget(data.response!, _sendRequest);
+        Widget buildChild() {
+          if (snapshot.hasData && snapshot.data != null) {
+            final data = snapshot.data!;
+            if (data.state == ReactiveState.loading) {
+              return widget.loadingWidget;
+            } else if (data.state == ReactiveState.failure) {
+              return widget.failureWidget(data.response!, _sendRequest);
+            } else if (data.state == ReactiveState.success) {
+              return widget.successWidget(data.response!, _sendRequest);
+            }
+          } else if (snapshot.hasError && snapshot.error != null) {
+            final error = snapshot.error! as ReactiveError;
+            return widget.retryWidget(error.exception, error.error, _sendRequest);
           }
-        } else if (snapshot.hasError && snapshot.error != null) {
-          print('hasError');
-          final error = snapshot.error! as ReactiveError;
-          return widget.retryWidget(error.exception, error.error, _sendRequest);
+
+          return widget.loadingWidget;
         }
 
-        print('unknown');
-        return widget.loadingWidget;
+        if (widget.wrapper != null && snapshot.hasData && snapshot.data != null) {
+          return widget.wrapper!(snapshot.data!.state, buildChild());
+        } else {
+          return buildChild();
+        }
       },
     );
+  }
+
+  Future<BaseResponse<DM>> _sendRequest() {
+    Completer<BaseResponse<DM>> completer = Completer<BaseResponse<DM>>();
+    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+      widget.request.send<DM>(
+        context,
+        onStart: () {
+          if (widget.listener != null) {
+            widget.listener!(ReactiveState.loading, _sendRequest);
+          }
+          _controller.add(ReactiveResponse(ReactiveState.loading));
+        },
+        onSuccess: (response) {
+          completer.complete(response);
+          if (response.isSuccessful) {
+            if (widget.listener != null) {
+              widget.listener!(ReactiveState.success, _sendRequest, response: response);
+            }
+            _controller.add(ReactiveResponse(ReactiveState.success, response: response));
+          } else {
+            if (widget.listener != null) {
+              widget.listener!(ReactiveState.failure, _sendRequest, response: response);
+            }
+            _controller.add(ReactiveResponse(ReactiveState.failure, response: response));
+          }
+        },
+        showRetry: false,
+        showProgress: false,
+        ignoreExpireTime: widget.ignoreExpireTime,
+        onError: (exception, error) {
+          completer.completeError(error);
+          final reactiveError = ReactiveError(
+            exception,
+            error,
+          );
+          if (widget.listener != null) {
+            widget.listener!(ReactiveState.error, _sendRequest, error: reactiveError);
+          }
+          _controller.addError(
+            reactiveError,
+          );
+        },
+      );
+    });
+    return completer.future;
   }
 
   @override
-  void onLoad(BuildContext context) {
-    _sendRequest();
-  }
-
-  void _sendRequest() {
-    widget.request.send<Res>(
-      context,
-      onStart: () {
-        print('sending request...');
-        if (widget.listener != null) {
-          widget.listener!(ReactiveState.loading, _sendRequest);
-        }
-        _controller.add(ReactiveResponse(ReactiveState.loading));
-      },
-      onSuccess: (response) {
-        print('response returned');
-        if (response.isSuccessful) {
-          if (widget.listener != null) {
-            widget.listener!(ReactiveState.success, _sendRequest, response: response);
-          }
-          _controller.add(ReactiveResponse(ReactiveState.success, response: response));
-        } else {
-          if (widget.listener != null) {
-            widget.listener!(ReactiveState.failure, _sendRequest, response: response);
-          }
-          _controller.add(ReactiveResponse(ReactiveState.failure, response: response));
-        }
-      },
-      showRetry: false,
-      showProgress: false,
-      ignoreExpireTime: widget.ignoreExpireTime,
-      onError: (exception, error) {
-        print('error occured');
-        final reactiveError = ReactiveError(
-          exception,
-          error,
-        );
-        if (widget.listener != null) {
-          widget.listener!(ReactiveState.error, _sendRequest, error: reactiveError);
-        }
-        _controller.addError(
-          reactiveError,
-        );
-      },
-    );
-  }
+  bool get wantKeepAlive => true;
 }