+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-## 1.0.1+3
+## 1.1.2
+
+* Added wrapper widget builder to ReactiveWidget
+
+## 1.1.1
+
+ReactiveWidget reload function now returns a future
+
+## 1.1.0
+
+Break Changes
+
+* Response class is removed
+* You can extend your Response models from DataModel class
+
+## 1.0.1+5
 
 * OnRetry
 * LoginBuilder